Delete nodejs.yml
Currently makes little sense to have this run without proper tests, only clogs resources and webhooks.
It does not make sense at this moment to run tests on an application that is dependent on the Discord API. This would require extra logins in a testing environment that would easily cause rate limits and other authentication issues.